Carry forward and carry last working, logging added (see note)
Carry forward and carry last both change how 'values' are obtained for accounts.
Carry forward allows values to 'extend' to a future date, while carry last
copys the 'last' value to a single specified date.

Invoking carry last modifies the account name so it's use is clear
in a report (although it will be hidden if included in a subsequent
meta-account)

Invoking carry-foward modifes how values can be obtained for arbitrary
periods, and thus how performance stats are calculated.  If this is invoked
get_performance returns a 'carry' key indicating the fact that can be displayed
in a report, etc.  A demo of this is currently shown in the BasicStatsReport