Improve tables/reports/ascii view (see note)
Previously tables/reports and the flattened view were a bit
of a mishmosh. Now:

Tables are rows of Cell instances each of
which has a string/float and object representation and
possibly some meta data.

Reports build tables, including giving a preferred justification
and width to each column

Views (ascii_view currently) take a report and present it to
the user this means producing a string representation of all
the Cells in a table along with their metadata